// utils/types.d.ts
export interface CarRental {
  id: string;
  title: string;
  price: string;
  location: string;
  images: string[];
  seater: number;
  packages: string[];
  isLiked?: boolean;
}